Moved LearnedDFL model to separate file
