+++ dfl/dfl.py
@@ -21,65 +21,20 @@ import copy
 
 import matplotlib.pyplot as plt
 
+from dfl.LearnedDFL import LearnedDFL
+
 np.set_printoptions(precision = 4)
 np.set_printoptions(suppress = True)
 
 H = 256
 dtype = torch.FloatTensor
-device = 'cpu'
+device = 'cpu' #'cuda' if torch.cuda.is_available() else 'cpu'
 seed = 9
 torch.manual_seed(seed)
 np.random.seed(seed = seed)
 
 RETRAIN = True
-# RETRAIN = False
-
-class LearnedDFL(torch.nn.Module):
-    def __init__(self, D_x, D_eta, D_u, H):
-        super(LearnedDFL, self).__init__()
-
-        self.D_x = D_x
-        D_xi = D_x + D_eta + D_u
-
-        self.g = torch.nn.Sequential(
-            torch.nn.Linear(D_x,H),
-            torch.nn.ReLU(),
-            torch.nn.ReLU(),
-            torch.nn.ReLU(),
-            torch.nn.Linear(H,D_eta)
-        )
-
-        self.A = torch.nn.Linear(D_xi,D_x)
-        self.H = torch.nn.Linear(D_xi,D_eta)
-
-    def forward(self, x_star):
-        x_tm1 = x_star[:,:self.D_x]
-        u_tm1 = x_star[:,self.D_x:]
-
-        eta_tm1 = self.g(x_tm1)
-
-        xi_tm1 = torch.cat((x_tm1,eta_tm1,u_tm1), 1)
-
-        x_t = self.A(xi_tm1)
-        eta_t = self.H(xi_tm1)
-
-        return x_t, eta_t
-
-    def lstsqAH(self, x_t, x_tm1, u_tm1):
-        eta_tm1 = self.g(x_tm1)
-
-        xi_tm1 = torch.cat((x_tm1,eta_tm1,u_tm1), 1)
-
-        eta_t = self.g(x_t)
-
-        A = torch.lstsq(  x_t,xi_tm1).solution[:5]
-        H = torch.lstsq(eta_t,xi_tm1).solution[:5]
-
-        for i in range(len(A)):
-            for j in range(len(A[0])):
-                self.A.weight[j,i] = A[i,j].item()
-            for j in range(len(H[0])):
-                self.H.weight[j,i] = H[i,j].item()
+RETRAIN = False
 
 def step(x_batch, y_batch, model, loss_fn):
     # Send data to GPU if applicable
@@ -190,7 +145,7 @@ class DFL():
 
             inp = torch.from_numpy(x).type(dtype)
 
-            eta = self.model.g(inp).detach().numpy().T.reshape(x_shape)
+            eta = self.model.g(inp).detach().numpy().T
 
             return eta
